src/requirements.txt
pystray==0.17.3
pillow==8.4.0
tk==0.1.0
